requirements.txt
flask==2.3.3
flask-cors==3.0.10
openai
pyjwt==2.1.0
python-dotenv==0.19.0
werkzeug==2.3.7